feat(790): wip struict
